Refactoring
Rename preferShortestDistanceToDistance to equalDistance as the estimate
function really doesn't care about any distance at all.

Rename dijkstra to astar, as it's in fact, an A* algrorithm. Now use a
equi-distance function to make a dijkstra and also build in a pattern
to correctly get the neighbours as we will need this for jumps